# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Edenlife Meal Recommendation System API V1 Documentation",
        "description": "Edenlife Meal Recommendation System API V1 Documentation",
        "contact": {
            "email": "ayodeleoniosun63@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/accounts/register": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Sign up a new account",
                "description": "Sign up a new account",
                "operationId": "29a92b4bc6a788f4c3b091747e71955f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email_address",
                                    "phone_number",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email_address": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "user": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "An error occured"
                    }
                }
            }
        },
        "/accounts/login": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Login to an existing account",
                "description": "Login to an existing account",
                "operationId": "62c068b998dd97929de262df320a4fbb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email_address",
                                    "password"
                                ],
                                "properties": {
                                    "email_address": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "user": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "An error occured"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}